classify/requirements.txt
azure-core==1.26.0
azure-cosmos==4.3.0
azure-storage-blob==12.9.0
click==8.0.3
pandas==1.3.2
python-dotenv==0.19.2
requests==2.26.0
setfit==0.6.0
dnspython==2.2.1
et-xmlfile==1.1.0
pandas==1.3.2
requests==2.26.0
xlsxwriter==3.0.2
python-dateutil==2.8.2
python-dotenv==0.19.2
openpyxl==3.0.9
pdfkit==1.0.0
azure-identity==1.10.0
azure-keyvault==4.2.0
pyodbc==4.0.34
pyzbar==0.1.9
greenlet==1.1.3.post0
eventlet==0.33.1
werkzeug==2.2.2
SQLAlchemy==1.4.41
lxml==4.9.1